+++ ml_tools/ensemble_inference.py
@@ -1,17 +1,16 @@
-from ._script_info import _script_info
-from ._logger import _LOGGER
-from .path_manager import make_fullpath, list_files_by_extension
-from .keys import EnsembleKeys
-
 from typing import Union, Literal, Dict, Any, Optional, List
 from pathlib import Path
 import json
-
 import joblib
 import numpy as np
 # Inference models
 import xgboost
 import lightgbm
+
+from ._script_info import _script_info
+from ._logger import _LOGGER
+from .path_manager import make_fullpath, list_files_by_extension
+from .keys import EnsembleKeys
 
 
 __all__ = [